+++ packages/documents/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tangleai/documents
 
+## 0.29.7
+
+### Patch Changes
+
+- @tangleai/context@0.29.7
+  - @tangleai/core@0.29.7
+  - @tangleai/models@0.29.7
+
 ## 0.29.6
 
 ### Patch Changes