+++ packages/pipeline/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tangleai/pipeline
 
+## 0.29.7
+
+### Patch Changes
+
+- @tangleai/core@0.29.7
+  - @tangleai/memory@0.29.7
+  - @tangleai/models@0.29.7
+
 ## 0.29.6
 
 ### Patch Changes