+++ packages/search/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tangleai/search
 
+## 0.29.7
+
+### Patch Changes
+
+- @tangleai/core@0.29.7
+
 ## 0.29.6
 
 ### Patch Changes